travelmanagementsystem/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
Deprecated==1.2.12
Django==3.2.6
django-ckeditor==6.1.0
django-debug-toolbar==3.2.1
django-js-asset==1.2.2
django-oauth-toolkit==1.5.0
djangorestframework==3.12.4
drf-yasg==1.20.0
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jwcrypto==1.0
MarkupSafe==2.0.1
mysqlclient==2.0.3
numpy==1.21.1
oauthlib==3.1.1
packaging==21.0
pandas==1.3.1
Pillow==8.3.1
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlclient==21.4.24
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
wrapt==1.12.1
